Refactored methods, responsible for SUBSCRIBE, UNSUBSCRIBE and TEAMINFO CallbackQueryHandlers. Add English locale messages, prepared structure for BotCommandParser structure.
