+++ src/main/java/com/vaoler/assistantcsgobot/bot/keyboards/handlers/callbackquery/CallbackQueryHandler.java
@@ -1,5 +1,6 @@
 package com.vaoler.assistantcsgobot.bot.keyboards.handlers.callbackquery;
 
+import com.vaoler.assistantcsgobot.bot.keyboards.handlers.callbackquery.types.CallbackQueryType;
 import org.telegram.telegrambots.meta.api.methods.send.SendMessage;
 import org.telegram.telegrambots.meta.api.objects.CallbackQuery;
