--- src/main/java/com/vaoler/assistantcsgobot/bot/keyboards/handlers/callbackquery/CallbackQueryType.java
@@ -1,4 +1,0 @@
-package com.vaoler.assistantcsgobot.bot.keyboards.handlers.callbackquery;
-
-public interface CallbackQueryType {
-}
