+++ src/main/java/com/vaoler/assistantcsgobot/bot/keyboards/handlers/callbackquery/types/CallbackQueryType.java
@@ -1,0 +1,4 @@
+package com.vaoler.assistantcsgobot.bot.keyboards.handlers.callbackquery.types;
+
+public interface CallbackQueryType {
+}
