+++ src/main/java/com/vaoler/assistantcsgobot/bot/keyboards/handlers/inputmessage/AssistantCSGOBotCommand.java
@@ -1,0 +1,4 @@
+package com.vaoler.assistantcsgobot.bot.keyboards.handlers.inputmessage;
+
+public enum AssistantCSGOBotCommand implements TelegramBotCommand{
+}
