+++ src/main/java/com/vaoler/assistantcsgobot/bot/keyboards/handlers/inputmessage/InputMessageHandler.java
@@ -1,8 +1,8 @@
-package com.vaoler.assistantcsgobot.bot.keyboards.handlers;
+package com.vaoler.assistantcsgobot.bot.keyboards.handlers.inputmessage;
 
 import org.telegram.telegrambots.meta.api.methods.send.SendMessage;
 import org.telegram.telegrambots.meta.api.objects.Message;
 
 public interface InputMessageHandler {
-    SendMessage handle(Message message);
+    SendMessage handleInputMessage(Message message);
 }
