+++ src/main/java/com/vaoler/assistantcsgobot/bot/keyboards/handlers/inputmessage/TelegramBotCommand.java
@@ -1,0 +1,4 @@
+package com.vaoler.assistantcsgobot.bot.keyboards.handlers.inputmessage;
+
+public interface TelegramBotCommand {
+}
