+++ src/main/resources/messages_en_EN.properties
@@ -1,16 +1,27 @@
 
 bot.commands.unknown_command = I don't know that command. Please, type "/" for display commands list. Or send message with text "help" to bot.
 
+# /start command messages
+bot.messages.start_command_message_new_user = Welcome! I am bot, which will help You to be in touch with your favourite team events! Type here how to use bot...
+bot.messages.start_command_message_old_user = Welcome back!
+
+# /help command messages
+bot.messages.help_message = Here will be the list of commands. but now ¯\_(?)_/¯
+
+# /mySubs command messages
+bot.messages.mySubs_emptySubsListMessage = You don't have favourite teams in your list. For subscribe to team send message to bot "/findTeam_Team name" and press Subscribe!
+bot.messages.mySubs_returnListOfSubs = Yours subscription list:
+
+# SUBSCRIBE callback query
+bot.message.callback.SUBSCRIBE.user_not_present = Something goes wrong. You are not in bot users list. Type to bot owner for get help.
+bot.message.callback.SUBSCRIBE.team_not_present = Something goes wrong. Team you are subscribed on is not present in teams list. Type to bot owner for get help.
+
+# callback UNSUBSCRIBE
+bot.message.callback.UNSUBSCRIBE.user_not_present = Something goes wrong. You are not in bot users list. Type to bot owner for get help.
+bot.message.callback.UNSUBSCRIBE.team_not_present = Something goes wrong. Team you are subscribed on is not present in teams list. Type to bot owner for get help.
+
 keyboard.greeting_message = Welcome! Choose section, please.
 inline.keyboard.teamsResults.greeting_message = Welcome! Choose the team, which results You want to see.\
                                                 Note, that results of last 5 days are displayed!
-
-inlineKeyboard.news.read = Read on HLTV.org
-inlineKeyboard.results.read = More info on HLTV.org
-
-reply.query.NEWS = You are jumped to News section
-reply.query.RESULTS = You are jumped to Matches Results section
-reply.query.MATCHES = You are jumped to Matches section
-reply.query.STATS = You are jumped to Match statistics section. For get statistic insert match id, please.
 
 reply.query.NOT_FOUND = Error! There is no such callback query type!
