+++ src/main/resources/messages_ru_RU.properties
@@ -12,6 +12,10 @@ bot.messages.help_message = Здесь будет список команд, а 
 bot.messages.mySubs_emptySubsListMessage = У Вас ещё нет избранных команд. Для того чтоб подписать наберите "/findTeam_Team name" и нажмите кнопку Subscribe
 bot.messages.mySubs_returnListOfSubs = Список избранных команд:
 
+# SUBSCRIBE callback query
+bot.message.callback.SUBSCRIBE.user_not_present = Что - то пошло не так. Вы отсутствуете в списке пользователей. Обратитесь к владельцу бота за помощью.
+bot.message.callback.SUBSCRIBE.team_not_present = Что - то пошло не так. Команда на которую Вы подписаны отсутствует в списке. Обратитесь к владельцу бота за помощью.
+
 # callback UNSUBSCRIBE
 bot.message.callback.UNSUBSCRIBE.user_not_present = Что - то пошло не так. Вы отсутствуете в списке пользователей. Обратитесь к владельцу бота за помощью.
 bot.message.callback.UNSUBSCRIBE.team_not_present = Что - то пошло не так. Команда на которую Вы подписаны отсутствует в списке. Обратитесь к владельцу бота за помощью.
